fastapi/requirements.txt
# # --------- Torch --------- #
-f https://download.pytorch.org/whl/torch_stable.html
torch==2.0.0+cpu
torchvision==0.15.1+cpu

# # --------- FastAPI --------- #
fastapi[all]==0.98.0
transformers==4.30.2
datasets==2.13.1
Pillow==9.5.0
gunicorn==20.1.0

# # --------- Others --------- #
tiktoken==0.4.0
gradio==3.39.0
pytest==7.3.2
loguru==0.7.0